amdea/requirements.txt
openai>=1.35.0
playwright>=1.44.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
httpx>=0.27.0
psutil>=5.9.0
pyautogui>=0.9.54
keyboard>=0.13.5
keyring>=25.2.0
jsonschema>=4.22.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.1
deepgram-sdk>=3.0.0
cartesia>=1.0.0
